Add `attempt` to formula macro to prevent errors from ruining state
The formula macro will attach update functions to any driving
references. If there is no guard on the update functions, it becomes
possible to change the values stored in the driving references that
error out when the formula runs.

Consider a formula with the body `(/ @a @b)`. When @b becomes 0, the
formula will fail (divide by zero), and with no guard, that error will
be run every time the driving references update, making the formula
'stuck'.

Instead, the update function is attempted in a try/catch that will
return `nil` on errors. This way, the formula itself is only ever
updated when the function succeeds, and the error doesn't block
anything.

Thus, if you create some invalid values in the driving references, you
can still recover the formula's proper functionality by reverting to
valid values in the references again.