Rework control block render fns into defmulti to allow extensions
Users can now define their own methods to render the control block as
well as the control-block-results.

Commonly, I suspect usage will revolve around modifying the
control-block-result renderers more than the block itself, but there
may be situations where both are necessasry. Consider, for example,
the X3D js library X_ITE, which provides custom html tags for 3D
content.

To make it work without constantly rebuilding the whole scene, the
results block must be wrapped inside a single x3d canvas. This way, on
re-render, the whole scene is not re-initialized, and content doesn't
have to reload, it can change dynamically. There may be other
situations where that sort of thing is necessary, but I suspsect a
much more common thing is a scenario as follows:

Imagine building a letcontrol block that builds up a map or seq of
maps, all with the same key structure. A user may wish to display
these in a table. Instead of creating a wrapping function inside the
letcontrol block, they could make a custom method to handle this and
just use the appropriate tag.