Derefs inside letcontrols body can drive updates. Hacky though.
This is one approach to allow derefs inside a letcontrols body drive
updates of that control's results. This means you can def a letcontrol
and use its value in other letcontrols and see the updates happen.

I don't completely love the solution because it creates an 'empty'
add-watch so that the driving refs can 'know' which control blocks
need to have their results sent to the frontend. So, every control
block formula adds a watch with the map `{:own-id :the-blocks-id}` so
that the ids can be found using `solenoid.utils/get-watches`. A bit
hacky!

I wanted to attach metadata to do this, but atoms don't allow metadata (at least not by
default) because they are not objects.