Implemented Toggle and (WIP) 2D point controls
Involved several changes:
- render-point-value fn in components
  - used also in server controller/:id route in a case that renders
  the value to be sent. Necessary for non-standard components (like
  the point control)
- EdnBlock uses textarea now
- default base component is better. All overrides are keywords in an
opts map, and defaults are always merged in so caller does not need to
know every value to provide, just their modified keys
- query-string parsing in utils now handles multiple of the same val,
which will be returned in a vector. This allows more complicated
shapes of data to be sent back. Necessary for the point control, and
probably useful for more complex controls too