Adding comments to try explain the dependents mechanism
It's a bit hard to follow what's happening without some comments to
try contextualize what's going on, so I added them here.

As it goes, I may still try to improve the explanations and design,
and probably will want to move things to a proper set of docs, and
link the more detailed explanation in docstrings instead.

But this works for now.