Derefs inside letcontrols body can drive updates. Hacky though. (#3)
* Derefs inside letcontrols body can drive updates. Hacky though.

This is one approach to allow derefs inside a letcontrols body drive
updates of that control's results. This means you can def a letcontrol
and use its value in other letcontrols and see the updates happen.

I don't completely love the solution because it creates an 'empty'
add-watch so that the driving refs can 'know' which control blocks
need to have their results sent to the frontend. So, every control
block formula adds a watch with the map `{:own-id :the-blocks-id}` so
that the ids can be found using `solenoid.utils/get-watches`. A bit
hacky!

I wanted to attach metadata to do this, but atoms don't allow metadata (at least not by
default) because they are not objects.

* get-derefs doesn't choke when there are maps in the body

I forgot that maps are seqable, so the destructuring in the mapped fn
breaks when a map is passed in, causing any letcontrol with any map in
the body to fail.

This predicate change to explicitly not include maps should fix that issue.

* Implemented Toggle and (WIP) 2D point controls

Involved several changes:
- render-point-value fn in components
  - used also in server controller/:id route in a case that renders
  the value to be sent. Necessary for non-standard components (like
  the point control)
- EdnBlock uses textarea now
- default base component is better. All overrides are keywords in an
opts map, and defaults are always merged in so caller does not need to
know every value to provide, just their modified keys
- query-string parsing in utils now handles multiple of the same val,
which will be returned in a vector. This allows more complicated
shapes of data to be sent back. Necessary for the point control, and
probably useful for more complex controls too

* Re-add accidentally dropped changes to make-hx-vals.

* Remove watches on controlblocks when those refs are removed

* Clean up the components a bit.

* Adding comments to try explain the dependents mechanism

It's a bit hard to follow what's happening without some comments to
try contextualize what's going on, so I added them here.

As it goes, I may still try to improve the explanations and design,
and probably will want to move things to a proper set of docs, and
link the more detailed explanation in docstrings instead.

But this works for now.

* Add ability to get all dependents and sub-dependents, with depth limit